refactor: 채팅 스크롤 관련 코드 useChatScroll hook으로 분리
